+++ CHANGELOG.md
@@ -7,8 +7,8 @@
 
 ### Embedded Sass
 
-* Fix a rare race condition in the embedded compiler that could be triggered by
-  a protocol error and cause a crash.
+* Fix a rare race condition where the embedded compiler could freeze when a
+  protocol error was immediately followed by another request.
 
 ## 1.68.0